Refactor: Move `DoraNode` implementation into submodule
